+++ javascript-modules/engines/jekyll-engine/lib/builder.js
@@ -1,0 +1,1 @@
+export const extensions = [".jekyll.html"];
