+++ javascript-modules/engines/jekyll-engine/lib/plugins/jsonify.js
@@ -1,7 +1,7 @@
 /**
  * Implementation of Jekyll's jsonify filter
  */
-module.exports = function (Liquid) {
+ export default function (Liquid) {
     this.registerFilter('jsonify', (classname) => {
         return JSON.stringify(classname);
     });
