+++ javascript-modules/engines/jekyll-engine/lib/plugins/slugify-plugin.js
@@ -1,10 +1,10 @@
-var slugify = require('slugify')
+import slugify from 'slugify';
 
 /**
  * Liquidjs implementation of Jekyll's slugify plugin
  * TODO: Behaviour might differ for complex strings
  */
-module.exports = function (Liquid) {
+ export default function (Liquid) {
     this.registerFilter('slugify', (text) => {
         if (text && typeof text === 'string') return slugify(text).toLowerCase()
         return text
